worker/requirements.txt
yfinance
openai
psycopg2-binary
python-dotenv
pandas
